pyproject.toml
#

# build-system
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

# project
[project]
authors = [{name = "Derek Wan", email = "d.wan@icloud.com"}]
dependencies = []
dynamic = ["version"]
name = "leet-code"
readme = "README.md"
requires-python = ">= 3.12"

[project.optional-dependencies]
dev = [
  "dycw-utilities[test] >= 0.26.64, < 0.27",
  "hatch >= 1.11.0, < 1.12",
  "ipython >= 8.24.0, < 8.25",
  "pytest-benchmark >= 4.0.0, < 4.1",
]

# coverage
[tool.coverage]

[tool.coverage.html]
directory = ".coverage/html"

[tool.coverage.report]
exclude_also = ["@overload", "if TYPE_CHECKING:"]
fail_under = 100.0
skip_covered = true
skip_empty = true

[tool.coverage.run]
branch = true
data_file = ".coverage/data"
parallel = true

# hatch
[tool.hatch]

[tool.hatch.build]
sources = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/leet_code"]

[tool.hatch.version]
path = "src/leet_code/__init__.py"

# nitpick
[tool.nitpick]
style = [
  "https://raw.githubusercontent.com/dycw/nitpick/master/styles/3.12.toml",
  "https://raw.githubusercontent.com/dycw/nitpick/master/styles/common.toml",
]

# pyright
[tool.pyright]
executionEnvironments = [{root = "src"}]
include = ["src"]
pythonVersion = "3.12"
reportImplicitOverride = "error"
reportImportCycles = "error"
reportMissingSuperCall = "error"
reportMissingTypeArgument = false
reportMissingTypeStubs = false
reportPrivateImportUsage = false
reportPrivateUsage = false
reportPropertyTypeMismatch = "error"
reportShadowedImports = "error"
reportUninitializedInstanceVariable = "error"
reportUnknownArgumentType = false
reportUnknownMemberType = false
reportUnknownParameterType = false
reportUnknownVariableType = false
reportUnnecessaryTypeIgnoreComment = "error"
reportUntypedBaseClass = false
reportUnusedCallResult = "error"
typeCheckingMode = "strict"

# pytest
[tool.pytest]

[tool.pytest.ini_options]
addopts = [
  "-ra",
  "-vv",
  "--color=auto",
  "--strict-markers",
]
filterwarnings = ["error"]
minversion = "8.0"
testpaths = ["src/tests"]
xfail_strict = true

# ruff
[tool.ruff]
extend-exclude = ["**/*.ipynb"]
src = ["src"]
target-version = "py312"
unsafe-fixes = true

[tool.ruff.lint]
ignore = [
  "ANN101",  # missing-type-self
  "ANN102",  # missing-type-cls
  "ANN401",  # dynamically-typed-expression
  "D",  # pydocstyle
  "E501",  # line-too-long
  "F403",  # undefined-local-with-import-star
  "F405",  # undefined-local-with-import-star-usage
  "PLE1142",  # await-outside-async
  "PLR0911",  # too-many-return-statements
  "PLR0913",  # too-many-arguments
  "PLR0915",  # too-many-statements
  "PLR2004",  # magic-value-comparison
  "PT013",  # pytest-incorrect-pytest-import
  "S101",  # assert-used
  "S603",  # subprocess-without-shell-equals-true
  "S607",  # start-process-with-partial-path
  # formatter
  "W191",  # tab-indentation
  "E111",  # indentation-with-invalid-multiple
  "E114",  # indentation-with-invalid-multiple-comment
  "E117",  # over-indented
  "D206",  # indent-with-spaces
  "D300",  # triple-single-quotes
  "Q000",  # bad-quotes-inline-string
  "Q001",  # bad-quotes-multiline-string
  "Q002",  # bad-quotes-docstring
  "Q003",  # avoidable-escaped-quote
  "COM812",  # missing-trailing-comma
  "COM819",  # prohibited-trailing-comma
  "ISC001",  # single-line-implicit-string-concatenation
  "ISC002",  # multi-line-implicit-string-concatenation
]
ignore-init-module-imports = false
select = ["ALL"]

[tool.ruff.lint.extend-per-file-ignores]
"src/tests/**/*.py" = [
  "S101",  # assert-used
]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.isort]
required-imports = ["from __future__ import annotations"]
